+++ packages/geo/src/lib/style/shared/feature/index.ts
@@ -1,0 +1,1 @@
+export * from './feature-style';
