+++ packages/geo/src/lib/style/shared/index.ts
@@ -1,3 +1,4 @@
 export * from './datasource';
 export * from './vector';
 export * from './overlay';
+export * from './feature';
